+++ run_api/user/events.py
@@ -7,20 +7,32 @@ from sqlalchemy import event
 from substrateinterface import Keypair
 from run_api.user.schemas import User
 from run_api.config import settings
+from run_api.utils import gen_random_token
+from run_api import constants
 
 
-@event.listens_for(User, "before_insert")
-def generate_uid(_, __, user):
+def generate_fingerprint(_, __, user: User):
+    """
+    Generate a fingerprint for the user.
+    Runs first
+    """
+    user.fingerprint = gen_random_token(k=16)
+
+
+def generate_uid(_, __, user: User):
     """
     Set the user_id based on hotkey.
+    Runs after fingerprint is generated.
     """
-    user.user_id = str(uuid.uuid5(uuid.NAMESPACE_OID, user.hotkey))
+    user.user_id = str(
+        uuid.uuid5(uuid.NAMESPACE_OID, user.fingerprint)
+    )  # NOTE: Change to fingerprint
 
 
-@event.listens_for(User, "before_insert")
-def generate_payment_address(_, __, user):
+def generate_payment_address(_, __, user: User):
     """
     Generate a new payment address for the user.
+    Runs AFTER user_id is generated.
     """
     mnemonic = Keypair.generate_mnemonic(words=24)
     keypair = Keypair.create_from_mnemonic(mnemonic)
@@ -29,3 +41,33 @@ def generate_payment_address(_, __, user):
         path=f"payments/tao/{user.user_id}",
         secret=dict(mnemonic=mnemonic, address=user.payment_address),
     )
+
+
+def ensure_hotkey(_, __, user: User):
+    """
+    Sets hotkey to zero address if not provided.
+    """
+    if not user.hotkey:
+        user.hotkey = constants.ZERO_ADDRESS_HOTKEY
+
+
+def ensure_coldkey(_, __, user: User):
+    """
+    Sets coldkey to payment address if not provided.
+    Runs AFTER payment address is generated.
+    """
+    if not user.coldkey:
+        user.coldkey = user.payment_address
+
+
+@event.listens_for(User, "before_insert")
+def handle_user_insert(_, __, user: User):
+    """
+    Handle user insert.
+    """
+    # Order is important here
+    generate_fingerprint(_, __, user)
+    generate_uid(_, __, user)
+    generate_payment_address(_, __, user)
+    ensure_coldkey(_, __, user)
+    ensure_hotkey(_, __, user)  # Order not important for this one