+++ run_api/user/schemas.py
@@ -11,6 +11,9 @@ from run_api.database import Base
 from run_api.config import settings
 
 
+# Other fields are populated by listeners
+# Except hotkey which is added in the header
+# NOTE: Can we add hotkey here instead?
 class UserRequest(BaseModel):
     username: str
     coldkey: str
@@ -19,12 +22,27 @@ class UserRequest(BaseModel):
 class User(Base):
     __tablename__ = "users"
 
+    # Populated in user/events based on fingerprint
     user_id = Column(String, primary_key=True)
+
+    # An alternative to an API key.
     hotkey = Column(String, nullable=False)
-    coldkey = Column(String, nullable=False)  # To receive commission payments
-    payment_address = Column(String)  # Users should send to this address to top up
-    balance = Column(BigInteger, default=settings.signup_bonus_balance)  # Current balance in Tao
-    username = Column(String, unique=True)  # Friendly name for the frontend for who created the chute
+
+    # To receive commission payments
+    coldkey = Column(String, nullable=False)
+
+    # Users should send to this address to top up
+    payment_address = Column(String)
+
+    # Current balance in Tao
+    balance = Column(BigInteger, default=settings.signup_bonus_balance)
+
+    # Friendly name for the frontend for chute creators
+    username = Column(String, unique=True)
+
+    # Gets populated in user/events.py to be a 16 digit alphanumeric which acts as an account id
+    fingerprint = Column(String, nullable=False, unique=True)
+    
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), server_default=func.now())
 
@@ -45,27 +63,28 @@ class User(Base):
             )
         return value
 
-    @validates("coldkey")
-    def validate_coldkey(self, _, value):
-        """
-        Ensure the coldkey has a balance before allowing registration.
-        """
-        if not re.match(r"^[a-zA-Z0-9]{48}$", value):
-            raise ValueError("Invalid coldkey address")
-        balance = 0
-        try:
-            substrate = SubstrateInterface(settings.subtensor, ss58_format=42)
-            result = substrate.query(
-                module="System", storage_function="Account", params=[value]
-            )
-            balance = float(result.value["data"]["free"]) / 1e9
-        except Exception as e:
-            raise ValueError(f"Error checking tao balance for {value}: {e}")
-        if not balance or balance < settings.registration_minimum_balance:
-            raise ValueError(
-                f"Free tao balance for coldkey={value} too low [{balance}], minimum: {settings.registration_minimum_balance}"
-            )
-        return value
+    # NOTE: The below can be deleted as coldkey is just for receiving payments.
+    # @validates("coldkey")
+    # def validate_coldkey(self, _, value):
+    #     """
+    #     Ensure the coldkey has a balance before allowing registration.
+    #     """
+    #     if not re.match(r"^[a-zA-Z0-9]{48}$", value):
+    #         raise ValueError("Invalid coldkey address")
+    #     balance = 0
+    #     try:
+    #         substrate = SubstrateInterface(settings.subtensor, ss58_format=42)
+    #         result = substrate.query(
+    #             module="System", storage_function="Account", params=[value]
+    #         )
+    #         balance = float(result.value["data"]["free"]) / 1e9
+    #     except Exception as e:
+    #         raise ValueError(f"Error checking tao balance for {value}: {e}")
+    #     if not balance or balance < settings.registration_minimum_balance:
+    #         raise ValueError(
+    #             f"Free tao balance for coldkey={value} too low [{balance}], minimum: {settings.registration_minimum_balance}"
+    #         )
+    #     return value
 
     def __repr__(self):
         """