+++ run_api/utils.py
@@ -3,6 +3,8 @@ Utility/helper functions.
 """
 
 import datetime
+import random
+import string
 import orjson as json
 
 
@@ -18,3 +20,10 @@ def sse(data):
     Format response object for server-side events stream.
     """
     return f"data: {json.dumps(data).decode()}\n\n"
+
+
+def gen_random_token(k: int = 16) -> str:
+    """
+    Generate a random token, useful for fingerprints.
+    """
+    return "".join(random.sample(string.ascii_letters + string.digits, k=k))